+++ lua/mini/statusline.lua
@@ -83,7 +83,8 @@
 --- >
 ---   function()
 ---     local mode, mode_hl = MiniStatusline.section_mode({ trunc_width = 120 })
----     local git           = MiniStatusline.section_git({ trunc_width = 75 })
+---     local git           = MiniStatusline.section_git({ trunc_width = 40 })
+---     local diff          = MiniStatusline.section_diff({ trunc_width = 75 })
 ---     local diagnostics   = MiniStatusline.section_diagnostics({ trunc_width = 75 })
 ---     local filename      = MiniStatusline.section_filename({ trunc_width = 140 })
 ---     local fileinfo      = MiniStatusline.section_fileinfo({ trunc_width = 120 })
@@ -92,7 +93,7 @@
 ---
 ---     return MiniStatusline.combine_groups({
 ---       { hl = mode_hl,                  strings = { mode } },
----       { hl = 'MiniStatuslineDevinfo',  strings = { git, diagnostics } },
+---       { hl = 'MiniStatuslineDevinfo',  strings = { git, diff, diagnostics } },
 ---       '%<', -- Mark general truncate point
 ---       { hl = 'MiniStatuslineFilename', strings = { filename } },
 ---       '%=', -- End left alignment
@@ -579,7 +580,8 @@ H.modes = setmetatable({
 H.default_content_active = function()
   -- stylua: ignore start
   local mode, mode_hl = MiniStatusline.section_mode({ trunc_width = 120 })
-  local git           = MiniStatusline.section_git({ trunc_width = 75 })
+  local git           = MiniStatusline.section_git({ trunc_width = 40 })
+  local diff          = MiniStatusline.section_diff({ trunc_width = 75 })
   local diagnostics   = MiniStatusline.section_diagnostics({ trunc_width = 75 })
   local filename      = MiniStatusline.section_filename({ trunc_width = 140 })
   local fileinfo      = MiniStatusline.section_fileinfo({ trunc_width = 120 })
@@ -591,7 +593,7 @@ H.default_content_active = function()
   -- sections, etc.)
   return MiniStatusline.combine_groups({
     { hl = mode_hl,                  strings = { mode } },
-    { hl = 'MiniStatuslineDevinfo',  strings = { git, diagnostics } },
+    { hl = 'MiniStatuslineDevinfo',  strings = { git, diff, diagnostics } },
     '%<', -- Mark general truncate point
     { hl = 'MiniStatuslineFilename', strings = { filename } },
     '%=', -- End left alignment